dynamic video parsing and split files
