+++ main.py
@@ -1,27 +1,22 @@
 # IMPORTS---------------------------------------------------------------------------------------------------------------
 import json
 from os import rename, remove
-import os.path
 import PIL.Image
 from PIL import Image, ImageFilter, ImageShow
 import PIL.Image
 import numpy as np
 import boto3
 import tkinter as tk
-from tkinter.filedialog import askopenfilename
-from tkinter import *
 from pathlib import Path
 import shutil
 import sys
 import random
 import csv
-import cv2
+import sqVideo
+import sqHeatmap
+import sqTimeline
 
 # CLASS DEFINITIONS-----------------------------------------------------------------------------------------------------
-class myVideo(object):
-
-    def __init__(self, file):
-        self.file = file
 
 # todo classes:
 #  image
@@ -31,40 +26,15 @@ class myVideo(object):
 
 # MAIN FUNCTION---------------------------------------------------------------------------------------------------------
 def main():
-    video = myVideo(selectVideo)
-    print("file: " + str(video.file))
+    # select video for use
+    video = sqVideo.myVideo(sqVideo.selectVideo())
+    print("file: " + str(video.fullpath))
 
-    parseVideo(video)
+    # parse video into images in a folder
+    sqVideo.parseVideo(video)
 
 
 # BODY FUNCTIONS--------------------------------------------------------------------------------------------------------
-
-# select video file from directory
-def selectVideo():
-    Tk().withdraw()
-    filename = askopenfilename()
-
-    return filename
-
-# parses the video into individual frames as images
-def parseVideo(video):
-    # todo:
-    #  create dynamic directory for images to save into
-    #  make video link dynamic
-
-    vid = cv2.VideoCapture("testVid.mp4")
-    success, image = vid.read()
-    count = 0
-    while success:
-        # saves every 5th frame of the video
-        if count % 5 == 0:
-            cv2.imwrite("testVid/" + "frame%d.jpg" % count, image)
-        success, image = vid.read()
-        count += 1
-
-    print("Video parsed")
-
-    return
 
 # todo functions:
 #  parse video to images
